# api/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Serverless Stable Diffusion Inference
  description: This is the serverless stable diffusion inference API.
paths:
  /models:
    post:
      summary: register model
      operationId: registerModel
      requestBody:
        description: model details
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelAttributes"
      responses:
        "200":
          description: register model response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: list model
      operationId: listModels
      responses:
        "200":
          description: list model all
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelAttributes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /models/{model_name}:
    put:
      summary: update model
      description: update model
      operationId: updateModel
      parameters:
        - name: model_name
          in: path
          description: name of model
          required: true
          schema:
            type: string
            example: "example_model_name"
      requestBody:
        description: model detail
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelAttributes"
      responses:
        "200":
          description: update model success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: delete model
      operationId: deleteModel
      parameters:
        - name: model_name
          in: path
          description: name of model
          required: true
          schema:
            type: string
            example: "example_model_name_to_delete"
      responses:
        "200":
          description: delete model success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: get model info
      operationId: getModel
      parameters:
        - name: model_name
          in: path
          description: name of model
          required: true
          schema:
            type: string
            example: "example_model_name_for_info"
      responses:
        "200":
          description: get model success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelAttributes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /txt2img:
    post:
      summary: txt to img predict
      operationId: txt2Img
      requestBody:
        description: predict params
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Txt2ImgRequest"
      responses:
        "200":
          description: submit predict success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitTaskResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /img2img:
    post:
      summary: img to img predict
      operationId: img2Img
      requestBody:
        description: predict params
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Img2ImgRequest"
      responses:
        "200":
          description: submit predict success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitTaskResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/progress:
    get:
      summary: get predict progress
      operationId: getTaskProgress
      parameters:
        - name: taskId
          in: path
          description: task id
          required: true
          schema:
            type: string
            example: "example_task_id_for_progress"
      responses:
        "200":
          description: get predict progress success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskProgressResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/cancellation:
    post:
      summary: cancel predict task
      operationId: cancelTask
      parameters:
        - name: taskId
          in: path
          description: task id
          required: true
          schema:
            type: string
            example: "example_task_id_to_cancel"
      responses:
        "200":
          description: cancel predict progress success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/result:
    get:
      summary: get predict result
      operationId: getTaskResult
      parameters:
        - name: taskId
          in: path
          description: task id
          required: true
          schema:
            type: string
            example: "example_task_id_for_result"
      responses:
        "200":
          description: get predict result success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResultResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /options:
    post:
      summary: update config options
      operationId: updateOptions
      requestBody:
        description: config option
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionRequest'
      responses:
        '200':
          description: update config options success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /extra_images:
    post:
      summary: image upcaling
      operationId: extraImages
      requestBody:
        description: image upcaling
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtraImagesRequest'
      responses:
        '200':
          description: image upcaling respone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitTaskResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restart:
    post:
      summary: restart webui api server
      operationId: restart
      responses:
        "200":
          description: restart success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Model:
      required:
        - type
        - name
        - ossPath
      properties:
        type:
          type: string
          description: model type
          example: "diffusion"
        name:
          type: string
          description: model name
          example: "model_v1"
        ossPath:
          type: string
          description: the oss path of the model
          example: "/path/to/oss/model_v1"
    ModelAttributes:
      allOf:
        - $ref: "#/components/schemas/Model"
        - required:
            - etag
            - status
          properties:
            etag:
              type: string
              description: the oss etag of the model
              example: "3f786850e387550fdab836ed7e6dc881de23001b"
            status:
              type: string
              description: the model status, registering, loading, loaded or unloaded
              example: "loaded"
            registeredTime:
              type: string
              description: the registered time of the model
              example: "2023-01-01T12:00:00Z"
            lastModificationTime:
              type: string
              description: the last modification time of the model
              example: "2023-01-10T12:00:00Z"
    Txt2ImgRequest:
      required:
        - stable_diffusion_model
        - sd_vae
        - prompt
      properties:
        stable_diffusion_model:
          type: string
          example: "diffusion_v1"
        sd_vae:
          type: string
          example: "vae_v1"
        save_dir:
          type: string
          example: "/path/to/save_dir"
        enable_hr:
          type: boolean
          example: true
        denoising_strength:
          type: number
          format: float
          example: 0.5
        firstphase_width:
          type: integer
          format: int32
          example: 640
        firstphase_height:
          type: integer
          format: int32
          example: 480
        hr_scale:
          type: integer
          format: int32
          example: 2
        hr_upscaler:
          type: string
          example: "upscale_method_v1"
        hr_second_pass_steps:
          type: integer
          format: int32
          example: 10
        hr_resize_x:
          type: integer
          format: int32
          example: 1280
        hr_resize_y:
          type: integer
          format: int32
          example: 960
        hr_sampler_name:
          type: string
          example: "sampler_v1"
        hr_prompt:
          type: string
          example: "High resolution prompt"
        hr_negative_prompt:
          type: string
          example: "Negative high resolution prompt"
        prompt:
          type: string
          example: "Mountain landscape during sunset"
        styles:
          type: array
          items:
            type: string
          example: ["style1", "style2"]
        seed:
          type: integer
          format: int32
          example: 123456
        subseed:
          type: integer
          format: int32
          example: 789
        subseed_strength:
          type: integer
          format: int32
          example: 5
        seed_resize_from_h:
          type: integer
          format: int32
          example: 480
        seed_resize_from_w:
          type: integer
          format: int32
          example: 640
        sampler_name:
          type: string
          example: "sampler_v2"
        batch_size:
          type: integer
          format: int32
          example: 32
        n_iter:
          type: integer
          format: int32
          example: 1000
        steps:
          type: integer
          format: int32
          example: 100
        cfg_scale:
          type: integer
          format: int32
          example: 1
        width:
          type: integer
          format: int32
          example: 640
        height:
          type: integer
          format: int32
          example: 480
        restore_faces:
          type: boolean
          example: true
        tiling:
          type: boolean
          example: false
        do_not_save_samples:
          type: boolean
          example: false
        do_not_save_grid:
          type: boolean
          example: false
        negative_prompt:
          type: string
          example: "Avoid mountains"
        eta:
          type: integer
          format: int32
          example: 5
        s_min_uncond:
          type: integer
          format: int32
          example: 1
        s_churn:
          type: integer
          format: int32
          example: 3
        s_tmax:
          type: integer
          format: int32
          example: 10
        s_tmin:
          type: integer
          format: int32
          example: 1
        s_noise:
          type: integer
          format: int32
          example: 2
        override_settings:
          type: object
          example: { "settingKey": "settingValue" }
        override_settings_restore_afterwards:
          type: boolean
          example: true
        script_args:
          type: array
          items:
            type: object
          example: [{ "argKey": "argValue" }]
        sampler_index:
          type: string
          example: "index_v1"
        script_name:
          type: string
          example: "script_v1"
        send_images:
          type: boolean
          example: true
        save_images:
          type: boolean
          example: true
        alwayson_scripts:
          type: object
          example: { "scriptKey": "scriptValue" }
...
    SubmitTaskResponse:
      required:
        - taskId
        - status
      properties:
        taskId:
          type: string
          example: "task123456"
        status:
          type: string
          example: "submitted"
        message:
          type: string
          example: "Task has been successfully submitted."
    TaskProgressResponse:
      required:
        - taskId
        - progress
        - etaRelative
        - currentImage
      properties:
        taskId:
          type: string
          example: "task123456"
        progress:
          type: number
          format: float
          example: 0.75
        etaRelative:
          type: number
          format: float
          example: 0.25
        state:
          type: object
          example: { "phase": "processing" }
        currentImage:
          type: string
          example: "/path/to/current/image.jpg"
        message:
          type: string
          example: "Processing image..."
...
    OptionRequest:
      description: config params
      required:
        - data
      properties:
        data:
          type: object
          example: "{}"
    ResponseMessage:
      description: response message
      required:
        - message
      properties:
        message:
          type: string
          example: "success"
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
          example: 400
        message:
          type: string
          description: Error message
          example: "Bad Request"